Add visualization tools for angle classification project
- Introduced multiple visualization scripts including `check_dataset.py`, `run_visualization.py`, and `visualize_individual_magnitudes.py` to facilitate dataset inspection and visualization of model outputs.
- Implemented `tsne_example.py` and `tsne_simple.py` for demonstrating t-SNE dimensionality reduction on synthetic datasets.
- Added `visualize_margin_loss.py` and `visualize_nn_outputs.py` for analyzing model performance and output distributions.
- Created `visualize_training.py` to plot training history and metrics.
- Updated `setup_imports.py` to streamline import statements across visualization scripts.
- Ensured all new scripts include necessary comments and documentation for clarity and maintainability.
