+++ visualizations/__init__.py
@@ -1,0 +1,12 @@
+"""
+Visualizations Package for Angle Classification Project
+
+This package contains various visualization utilities for the angle classification project.
+"""
+
+# Add the parent directory to the Python path
+import os
+import sys
+parent_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+if parent_dir not in sys.path:
+    sys.path.insert(0, parent_dir)
